move modals folder from common to components
