+++ packages/frinx-workflow-ui/src/common/components/modals/definition-modal.tsx
@@ -9,7 +9,7 @@ import {
   ModalHeader,
   ModalOverlay,
 } from '@chakra-ui/react';
-import { ClientWorkflow, Editor } from '@frinx/shared/src';
+import { ClientWorkflow, Editor } from 'packages/shared/src';
 
 type ModalProps = {
   isOpen: boolean;
