+++ packages/frinx-workflow-ui/src/common/components/modals/edit-schedule-workflow-modal.tsx
@@ -20,7 +20,6 @@ import {
   ModalFooter,
   ModalHeader,
   ModalOverlay,
-  Progress,
   Spacer,
   Text,
   useBoolean,
@@ -31,64 +30,34 @@ import {
   ClientWorkflow,
   getInitialValuesFromParsedInputParameters,
   CreateScheduledWorkflow,
-} from '@frinx/shared/src';
+  EditScheduledWorkflow,
+} from 'packages/shared/src';
 import { useFormik } from 'formik';
 import * as Yup from 'yup';
 import moment from 'moment';
-import ExecuteWorkflowModalFormInput from '@frinx/shared/src/components/execute-workflow-modal/execute-workflow-modal-form-input';
+import ExecuteWorkflowModalFormInput from 'packages/shared/src/components/execute-workflow-modal/execute-workflow-modal-form-input';
 import FeatherIcon from 'feather-icons-react';
-import { gql, useQuery } from 'urql';
-import { SchedulesQuery, SchedulesQueryVariables } from '../../__generated__/graphql';
-
-const SCHEDULED_WORKFLOWS_QUERY = gql`
-  query GetSchedules {
-    schedules {
-      edges {
-        node {
-          name
-        }
-      }
-    }
-  }
-`;
 
 const DEFAULT_CRON_STRING = '* * * * *';
 const CRON_REGEX = /^(\*|[0-5]?\d)(\s(\*|[01]?\d|2[0-3])){2}(\s(\*|[1-9]|[12]\d|3[01])){2}$/;
 
 type Props = {
+  scheduledWorkflow: EditScheduledWorkflow;
   workflow: ClientWorkflow;
   isOpen: boolean;
   onClose: () => void;
-  onSubmit: (workflow: CreateScheduledWorkflow) => void;
+  onSubmit: (workflow: EditScheduledWorkflow) => void;
 };
 
-const ScheduleWorkflowModal: FC<Props> = ({ workflow, isOpen, onClose, onSubmit }) => {
-  const [{ data: scheduledWorkflows, fetching: isLoadingScheduledWorkflows }] = useQuery<
-    SchedulesQuery,
-    SchedulesQueryVariables
-  >({
-    query: SCHEDULED_WORKFLOWS_QUERY,
-  });
+const resetDateFormat = (dateTime: string | undefined): string => {
+  return moment(dateTime).format('yyyy-MM-DDTHH:mm');
+};
 
+const ScheduleWorkflowModal: FC<Props> = ({ scheduledWorkflow, workflow, isOpen, onClose, onSubmit }) => {
   const validationSchema = Yup.object().shape({
     workflowName: Yup.string().required('Workflow name is required'),
     workflowVersion: Yup.number().required('Workflow version is required'),
-    name: Yup.string()
-      .required('Schedule name is required')
-      .test({
-        name: 'name',
-        message: 'Schedule name must be unique',
-        test: (value) => {
-          const scheduleNames = scheduledWorkflows?.schedules.edges.map((edge) => {
-            return edge?.node.name;
-          });
-          const isNameUnique = !scheduleNames?.some((wfName) => wfName === value);
-          if (!isNameUnique) {
-            return false;
-          }
-          return true;
-        },
-      }),
+    name: Yup.string().required('Schedule name is required'),
     cronString: Yup.string()
       .required('Cron expression is required')
       .test({
@@ -118,18 +87,21 @@ const ScheduleWorkflowModal: FC<Props> = ({ workflow, isOpen, onClose, onSubmit 
     validationSchema,
     validateOnMount: false,
     initialValues: {
-      workflowName: workflow.name,
-      workflowVersion: workflow.version?.toString() ?? '',
-      workflowContext: getInitialValuesFromParsedInputParameters(parsedInputParameters, dynamicInputParameters),
-      name: '',
-      cronString: DEFAULT_CRON_STRING,
-      isEnabled: false,
-      performFromDate: undefined,
-      performTillDate: undefined,
+      workflowName: scheduledWorkflow?.workflowName ?? workflow.name,
+      workflowVersion: scheduledWorkflow?.workflowVersion ?? workflow.version?.toString() ?? '',
+      workflowContext: scheduledWorkflow?.workflowContext
+        ? scheduledWorkflow.workflowContext
+        : getInitialValuesFromParsedInputParameters(parsedInputParameters, dynamicInputParameters),
+      name: scheduledWorkflow?.name || '',
+      cronString: scheduledWorkflow?.cronString || DEFAULT_CRON_STRING,
+      isEnabled: scheduledWorkflow?.isEnabled ?? false,
+      performFromDate: scheduledWorkflow?.performFromDate && resetDateFormat(scheduledWorkflow?.performFromDate),
+      performTillDate: scheduledWorkflow?.performFromDate && resetDateFormat(scheduledWorkflow?.performTillDate),
     },
     onSubmit: (formValues) => {
       const formattedValues = {
         ...formValues,
+        id: scheduledWorkflow?.id,
         cronString: formValues.cronString || DEFAULT_CRON_STRING,
         ...(formValues.performFromDate && {
           performFromDate: moment(formValues.performFromDate).format('yyyy-MM-DDTHH:mm:ss.SSSZ'),
@@ -141,6 +113,7 @@ const ScheduleWorkflowModal: FC<Props> = ({ workflow, isOpen, onClose, onSubmit 
           workflowContext: formValues.workflowContext,
         }),
       };
+
       onSubmit(formattedValues);
       onClose();
     },
@@ -156,10 +129,6 @@ const ScheduleWorkflowModal: FC<Props> = ({ workflow, isOpen, onClose, onSubmit 
     );
   };
 
-  if (isLoadingScheduledWorkflows) {
-    return <Progress size="xs" isIndeterminate />;
-  }
-
   const inputParametersKeys = Object.keys(values.workflowContext);
 
   return (
@@ -174,6 +143,7 @@ const ScheduleWorkflowModal: FC<Props> = ({ workflow, isOpen, onClose, onSubmit 
           <FormControl isInvalid={errors.name != null} isRequired>
             <FormLabel>Schedule name</FormLabel>
             <Input
+              disabled={scheduledWorkflow?.name.length !== undefined}
               value={values.name}
               onChange={(e) => {
                 handleChange(e);
